Sanitise params used off the URL path
Its possible pass params that end up being used with File.read/write
methods such as '..' which could allow access to files outside of the
filestore root. I haven't been able to do this with my testing as
it either doesn't parse the path correctly of the Rack middleware blows
up. Either way, this patch ensures that params match only alpha-numeric
as well as - and _ can be used.

This project is only intended to be used on networks where the network
itself provides the security but its still probably worth-while adding
this in.